Rename Config to Hooks
